+++ relay-server/src/metrics_extraction/event.rs
@@ -1130,7 +1130,7 @@ mod tests {
                     "timestamp": 1702474613.0495,
                     "start_timestamp": 1702474613.0175,
                     "description": "Autofix Pipeline",
-                    "op": "ai.pipeline.langchain",
+                    "op": "ai.pipeline",
                     "origin": "auto.langchain",
                     "span_id": "9c01bd820a083e63",
                     "parent_span_id": "a1e13f3f06239d69",
@@ -1534,7 +1534,9 @@ mod tests {
 
         assert!(!metrics.is_empty());
         for metric in metrics {
-            if &*metric.name == "d:spans/exclusive_time@millisecond" {
+            if &*metric.name == "d:spans/exclusive_time@millisecond"
+                || &*metric.name == "d:spans/duration@millisecond"
+            {
                 assert_eq!(metric.tag("ttid"), Some("ttid"));
                 assert_eq!(metric.tag("ttfd"), Some("ttfd"));
             } else {