+++ relay-server/src/metrics_extraction/sessions/mod.rs
@@ -493,6 +493,9 @@ mod tests {
                 },
                 metadata: BucketMetadata {
                     merges: 1,
+                    received_at: Some(
+                        UnixTimestamp(0),
+                    ),
                 },
             },
             Bucket {
@@ -512,6 +515,9 @@ mod tests {
                 },
                 metadata: BucketMetadata {
                     merges: 1,
+                    received_at: Some(
+                        UnixTimestamp(0),
+                    ),
                 },
             },
             Bucket {
@@ -531,6 +537,9 @@ mod tests {
                 },
                 metadata: BucketMetadata {
                     merges: 1,
+                    received_at: Some(
+                        UnixTimestamp(0),
+                    ),
                 },
             },
             Bucket {
@@ -550,6 +559,9 @@ mod tests {
                 },
                 metadata: BucketMetadata {
                     merges: 1,
+                    received_at: Some(
+                        UnixTimestamp(0),
+                    ),
                 },
             },
             Bucket {
@@ -569,6 +581,9 @@ mod tests {
                 },
                 metadata: BucketMetadata {
                     merges: 1,
+                    received_at: Some(
+                        UnixTimestamp(0),
+                    ),
                 },
             },
             Bucket {
@@ -588,6 +603,9 @@ mod tests {
                 },
                 metadata: BucketMetadata {
                     merges: 1,
+                    received_at: Some(
+                        UnixTimestamp(0),
+                    ),
                 },
             },
             Bucket {
@@ -609,6 +627,9 @@ mod tests {
                 },
                 metadata: BucketMetadata {
                     merges: 1,
+                    received_at: Some(
+                        UnixTimestamp(0),
+                    ),
                 },
             },
         ]